Refactors to introduce DraftServiceDAO and DraftFileStorageServiceDAO so that two level mocks aren't needed for GCSService
